requirements.txt
# wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
# bash Miniconda3-latest-Linux-x86_64.sh
# cd nsql
# conda create -n NSQL python=3.11 -y
# conda activate NSQL
# curl -fsSL https://ollama.com/install.sh | sh
# nohup ollama serve > ollama.log 2>&1 &
# ollama pull mistral
# ollama pull prompt/nsql-7b
# pip install -r requirements.txt

# Empty dataframe removal
# constraints as a prompt
# rag refers to currency table for which the credit card operates
# validation

pandas
SQLAlchemy
psycopg2-binary
pandasai
langchain_community
ollama